tags/1.0.0/requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.2
django-admin-list-filter-dropdown==1.0.1
django-admin-sortable2==0.6.13
django-autocomplete-light==3.2.9
django-ckeditor==5.2.2
django-debug-toolbar==1.8
django-flat-theme==1.1.4
django-fluent-pages==1.1.1
django-fluent-utils==1.3.1
django-jsoneditor==0.0.11
django-modeladmin-reorder==0.2
django-mptt==0.8.7
django-parler==1.8
django-polymorphic==1.2
django-polymorphic-tree==1.4
django-slug-preview==1.0.2
django-tag-parser==3.1
django-timezone-field==2.0
et-xmlfile==1.0.1
future==0.16.0
idna==2.5
jdcal==1.3
mysqlclient==1.3.10
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
packaging==16.8
Pillow==4.1.1
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytidylib==0.3.2
pytz==2017.2
pyup-django==0.4.0
PyYAML==3.12
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
tablib==0.11.5
unicodecsv==0.14.1
urllib3==1.21.1
xlrd==1.0.0
xlwt==1.2.0
